fix: 敵関連のリファクタ
