+++ Assets/Soroeru/Prefabs/Enemy/Enemy Variant - Normal.prefab
@@ -82,7 +82,7 @@ GameObject:
   m_CorrespondingSourceObject: {fileID: 5856298970032762474, guid: a13dfb45fa7b34c4a96385398dd8cacf, type: 3}
   m_PrefabInstance: {fileID: 2789556165366022656}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &6780796507271058425
+--- !u!114 &1797856677597482353
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -91,7 +91,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 8643461497611971690}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2c829d1c6baa4b1ba91c0ca9ee6085c8, type: 3}
+  m_Script: {fileID: 11500000, guid: 99a71e54f4c742fcbdec39a141c5fd98, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   attackPower: 3
