+++ Assets/Soroeru/Scripts/InGame/Presentation/View/UniView.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 99a71e54f4c742fcbdec39a141c5fd98
+timeCreated: 1651830107
